# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yachan-v2 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/{board_slug}/posts/{post_number}": {
      "patch": {
        "tags": [
          "posts"
        ],
        "summary": "Edit Post",
        "operationId": "edit_post_api__board_slug__posts__post_number__patch",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "post_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEditCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mod/{board_slug}/posts/{post_number}": {
      "delete": {
        "tags": [
          "mod"
        ],
        "summary": "Delete Post",
        "operationId": "delete_post_api_mod__board_slug__posts__post_number__delete",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "post_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "media_type",
          "original_name",
          "url",
          "mime_type"
        ],
        "title": "AttachmentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MediaType": {
        "type": "string",
        "enum": [
          "image",
          "video",
          "gif"
        ],
        "title": "MediaType"
      },
...
      "PostEditCreate": {
        "properties": {
          "new_body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "New Body"
          }
        },
        "type": "object",
        "title": "PostEditCreate"
      },
...
      "PostResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "post_number": {
            "type": "integer",
            "title": "Post Number"
          },
          "thread_id": {
            "type": "integer",
            "title": "Thread Id"
          },
          "board_id": {
            "type": "integer",
            "title": "Board Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tripcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripcode"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "body_html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Html"
          },
          "sage": {
            "type": "boolean",
            "title": "Sage"
          },
          "is_op": {
            "type": "boolean",
            "title": "Is Op"
          },
          "is_edited": {
            "type": "boolean",
            "title": "Is Edited"
          },
          "edited_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "post_number",
          "thread_id",
          "board_id",
          "name",
          "tripcode",
          "body",
          "body_html",
          "sage",
          "is_op",
          "is_edited",
          "edited_at",
          "created_at"
        ],
        "title": "PostResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}